fix(typings): hide injected props from the HOC result
In the current typings, the injected props (e.g. coords) protruded into the props of the resulting

component, and therefore polluted the code completion with invalid values. This is fixed now only

the 'non-Geolocated' props remain  in the resulting components' props.

BREAKING CHANGE: Injected props are no longer available for setting in the resulting components

#113